Split up fancy types to "events.pyi"
